simplify

+++ src/h3/ffi.rs
@@ -71,14 +71,14 @@ pub extern fn quiche_h3_config_set_qpack_blocked_streams(
 
 #[no_mangle]
 pub extern fn quiche_h3_config_set_dgram_poll_threshold(
-    config: &mut h3::Config, v: usize,
+    config: &mut h3::Config, v: u64,
 ) {
     config.set_dgram_poll_threshold(v);
 }
 
 #[no_mangle]
 pub extern fn quiche_h3_config_set_stream_poll_threshold(
-    config: &mut h3::Config, v: usize,
+    config: &mut h3::Config, v: u64,
 ) {
     config.set_stream_poll_threshold(v);
 }
